\name{tensorMQR1-package}
\alias{tensorMQR1-package}
\alias{tensorMQR1}
\docType{package}
\title{
Symmetric Tensor Estimation for Quadratic Regression.
}
\description{
  For a high-dimensional multiresponse quadratic regression (MQR) with or without aparsity assumptions, 
  treating the coefficients as a third-order tensor and borrowing Tucker decomposition to reduce the number of parameters.  
  The multivariate sparse group lasso (mcp or scad) and the steepest gradient descent algorithm are used to estimate
  tensor for sparsity situation.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Xu Liu

Maintainer: Xu Liu <liu.xu@sufe.edu.cn>
}
\references{
Symmetric Tensor Estimation for Quadratic Regression.
}
\keyword{High-dimensional; Gradient descent; Stiefel manifold;  Tensor estimation; Tucker decomposition. }
